factors use long

+++ src/Factors.java
@@ -1,9 +1,9 @@
 void main() {
     print("enter a number ");
     Scanner sc = Tools.scanner;
-    int a = sc.nextInt();
-    int b = 1;
-    int c = 0;
+    long a = sc.nextLong();
+    long b = 1;
+    long c = 0;
     while(b <= a) {
         if(a % b == 0) {
             print(b + " ");
